ocr_backend/requirements.txt
Flask==3.1.1
Flask-Cors==5.0.1
gunicorn==23.0.0
Pillow==11.3.0
paddleocr==2.9.1
pytest==8.4.1
